Switch class hierarchy [skip ci]
